Laravel: moved Http Kernel hooks.
