+++ src/Instrumentation/Laravel/src/Hooks/Illuminate/Foundation/Application.php
@@ -13,7 +13,6 @@ use OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers\ClientRequestWatcher;
 use OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers\ExceptionWatcher;
 use OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers\LogWatcher;
 use OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers\QueryWatcher;
-use OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers\RequestWatcher;
 use OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers\Watcher;
 use function OpenTelemetry\Instrumentation\hook;
 use Throwable;
@@ -33,7 +32,6 @@ class Application implements LaravelHook
                 $this->registerWatchers($application, new ExceptionWatcher());
                 $this->registerWatchers($application, new LogWatcher());
                 $this->registerWatchers($application, new QueryWatcher($this->instrumentation));
-                $this->registerWatchers($application, new RequestWatcher());
             },
         );
     }
