+++ src/Instrumentation/Laravel/src/Propagators/ResponsePropagationSetter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Contrib\Instrumentation\Laravel;
+namespace OpenTelemetry\Contrib\Instrumentation\Laravel\Propagators;
 
 use function assert;
 use Illuminate\Http\Response;
